refactor: Directory 생성 로직 간소화 및 Validator 분리
